Refactor CORS handling by moving cors.php to bots directory and updating include paths in index.php, token.php, and include.php
