+++ bots/access_helps.php
@@ -7,7 +7,7 @@ use function Publish\AccessHelps\get_access_from_db;
 use function Publish\AccessHelps\del_access_from_db;
 */
 
-include_once __DIR__ . '/include.php';
+include_once __DIR__ . '/../include.php';
 
 use function Publish\MdwikiSql\execute_query;
 use function Publish\MdwikiSql\fetch_query;
