+++ bots/cors.php
