+++ bots/do_edit.php
@@ -3,12 +3,11 @@
 namespace Publish\DoEdit;
 /*
 Usage:
-include_once __DIR__ . '/do_edit.php';
 use function Publish\DoEdit\publish_do_edit;
 use function Publish\DoEdit\get_edits_token;
 */
 
-include_once __DIR__ . '/include.php';
+include_once __DIR__ . '/../include.php';
 
 use MediaWiki\OAuthClient\Client;
 use MediaWiki\OAuthClient\ClientConfig;
