+++ include.php
@@ -9,15 +9,16 @@ include_once __DIR__ . '/../auth/vendor_load.php';
 
 include_once __DIR__ . '/bots/mdwiki_sql.php';
 
-include_once __DIR__ . '/config.php';
-include_once __DIR__ . '/helps.php';
-include_once __DIR__ . '/access_helps.php';
-include_once __DIR__ . '/access_helps_new.php';
-include_once __DIR__ . '/do_edit.php';
-include_once __DIR__ . '/add_to_db.php';
-include_once __DIR__ . '/get_token.php';
+include_once __DIR__ . '/bots/config.php';
+include_once __DIR__ . '/bots/helps.php';
+include_once __DIR__ . '/bots/access_helps.php';
+include_once __DIR__ . '/bots/access_helps_new.php';
+include_once __DIR__ . '/bots/do_edit.php';
+include_once __DIR__ . '/bots/add_to_db.php';
+include_once __DIR__ . '/bots/get_token.php';
 include_once __DIR__ . '/textfixes/include.php';
 include_once __DIR__ . '/bots/wd.php';
+include_once __DIR__ . '/bots/process_edit.php';
 
 $fix_refs_file = __DIR__ . '/fix_refs/work.php';
 $fix_refs_file2 = __DIR__ . '/../fix_refs_repo/work.php';
